Derive the status bar, and add GM mode and bank payment
Four fields of the 0x11 status bar were constants — gold 0, armour 0, weight a
flat body weight, followers 0 — read by a player every session with no way to
check them. They are read-site derivations now, the shape `equipped_weapon`
established: `items::carried` walks the worn tree once for both the gold total
and the carried weight (gold at ServUO's 0.02 stones a coin, everything else at
its tiledata weight through a new `Terrain::item_weight`), armour from a new core
table keyed by graphic (`combat::armor`, the classic suits and shields from
ServUO's `BaseArmor`, with the layer a piece is worn on supplying where it sits),
followers from whether a mount is under the rider. Two edges are
`Mobile.UpdateTotals`' exactly: the bank box is not carried (it is
`IsVirtualItem`, skipped for weight and gold alike) and a held item is
(`m_Holding` is added explicitly). The bar re-sends by diffing twice a second
over the online players (`tick/status.rs`) rather than by a call beside every
item mutation, which is the pattern the persistence rule warns decays.

The armour rating is not decorative: pre-AoS a swing gives up a share of it
through `absorb_physical`, ServUO's `BaseWeapon.AbsorbDamage` — a rolled hit
location, that piece and any shield eating their own bite, then a cut of the
wearer's total, all on the seeded rng. Where ServUO's two ladders disagree (its
piece selection gives the arms the 14%-wide band while its `ArmorScalars` array
gives the helm 0.14) the scalars array wins, and a test holds the two together.

Hit points regenerate (`combat::regen_hits`, ServUO's eleven-second pre-AoS rate,
none while dead or poisoned) — mana and stamina trickled back and hits never did.
And the stamina pool gains its first consumer: `combat::spend_step_stamina` is
ServUO's `WeightOverloading`, so an overloaded step costs 5 + over/25 (a third
mounted, double running), a nearly empty pool costs an extra point, every
sixteenth step on foot costs one anyway, and a pool at zero refuses the step
through the same 0x21 path paralysis uses.

Testing any of that from a staff account was impossible, because staff are exempt
from every such rule. So Sphere's split lands too: `PLEVEL` (the account's
authority) and `PRIV_GM` (the flag that exempts a character) are two things, and
`.GM` toggles only the second. `Access` stays the authority and a new `Staff`
marker is the mode, given at login to a game master and taken off by `.gm`;
`is_staff` reads the marker (fatigue, seeing and hearing the dead) and a new
`staff_authority` reads the level (every command gate), or `.gm off` would lock a
game master out of `.gm on`.

Moving the bank box out of what a character carries left banked gold buying
nothing, so two `[gameplay]` flags put it back: `vendor_bank_payment` (default
on, ServUO's `BaseVendor` — the pack first, then the bank, whole, and the vendor
says which paid) and `bank_gold_in_status` (default off, ServUO's truth). Weight
is on neither. `items::banked_gold` is the one counter for all three readers and
walks the whole box, so a purse inside it counts where the old one-level scan
missed it; `take_from_container` widened to u32 with it, since a bank purchase
runs past one u16 stack and the old cap refused it with "thou canst not afford
that".

Two smaller truths ride along. The client's 0xD1 logout is answered
(`encode_logout_ack`, both references' two bytes) — the id sat in the length
table with nothing routing it, so the paperdoll's Log Out hung until the client
timed out; it is also the one entry where the references disagree on the incoming
length, and the table now takes ServUO's two with the argument written beside it.
And the living no longer hear the dead: a ghost was drawn only to ghosts and
staff but audible to everyone, so `chat::speak` filters its listeners through the
same `can_see_mobile` that gates drawing.

Also records in the roadmap the gaps a sweep of it turned up that were never
written down: regions, day/night and weather, fame and karma, crafting and
harvesting, party and chat channels, pets and taming, and CI.

+++ crates/magic/src/lib.rs
@@ -196,7 +196,7 @@ fn has_reagents(state: &WorldState, pack: u32, reagents: &[(u16, u16)]) -> bool 
 fn consume_reagents(state: &mut WorldState, pack: u32, reagents: &[(u16, u16)]) {
     if let Some(pack) = Serial::new(pack) {
         for &(graphic, count) in reagents {
-            take_from_container(state, pack, graphic, count);
+            take_from_container(state, pack, graphic, u32::from(count));
         }
     }
 }
